Fixed type error causing infinite loop when a was too large

+++ bsp_prototype.c
@@ -448,7 +448,7 @@ void run_workload(int w, gsl_rng *r, double a, double b, double cpn)
   case WORKLOAD_FWQ:
     inter_time = generate_interval_rng(r, rng_type, a, b);
     assert(inter_time >= 0.0);
-    for (int i = 0; i < FWQ_CALIBRATE * inter_time; i++) {
+    for (int i = 0; i < (int)((int)FWQ_CALIBRATE * (int)inter_time); i++) {
       WORKLOAD_VALUE += i;
       asm("");
     }
